+++ librad/src/git/storage.rs
@@ -38,9 +38,9 @@ use crate::{
             Oid,
             References,
         },
+        p2p::url::{GitUrl, GitUrlRef},
         refs::{self, Refs},
         repo::Repo,
-        transport::p2p::{GitUrl, GitUrlRef},
         types::Reference,
     },
     hash::Hash,
