+++ librad/src/git/transport.rs
@@ -16,4 +16,3 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 pub mod local;
-pub mod p2p;
