+++ librad/src/net/peer.rs
@@ -28,7 +28,7 @@ use thiserror::Error;
 use crate::{
     git::{
         self,
-        server::GitServer,
+        p2p::server::GitServer,
         storage::{self, Storage as GitStorage, WithSigner},
     },
     internal::{borrow::TryToOwned, channel::Fanout},
@@ -264,7 +264,7 @@ impl Peer {
         );
 
         let protocol = Protocol::new(gossip, git);
-        git::transport::p2p::register()
+        git::p2p::transport::register()
             .register_stream_factory(&peer_id, Box::new(protocol.clone()));
 
         let run_loop = protocol
